+++ config/routes.rb
@@ -109,6 +109,7 @@ Rails.application.routes.draw do
         patch :update_listing
         post :publish_to_etsy
         post :relist_on_etsy
+        post :push_video_to_etsy
         post :regenerate_listing_images
         post :regenerate_listing_video
         post :upload_listing_video